Refactor components by functionality rather than phase
