+++ src/components/Button/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
-interface IProps {
+interface Props {
   children: React.ReactNode;
   onClick: () => void;
   disabled?: boolean;
   type?: 'button' | 'submit' | 'reset';
 }
 
-export const Button: React.FC<IProps> = (props) => {
+export const Button: React.FC<Props> = (props) => {
   return (
     <button onClick={props.onClick} disabled={props.disabled} type={props.type}>
       {props.children}
