+++ src/components/ViewController/connector/index.ts
@@ -4,12 +4,12 @@ import { IState } from '../../../store/reducers';
 import { PHASE_NAME } from '../../../types/phase';
 import { PLAYER_ROLE } from '../../../types/player';
 
-export interface IPropsFromState {
+export interface PropsFromState {
   phaseName?: PHASE_NAME;
   role?: PLAYER_ROLE;
 }
 
-export const mapStateToProps = (state: IState): IPropsFromState => {
+export const mapStateToProps = (state: IState): PropsFromState => {
   const self = state.game?.players.find(
     (player) => player.name === state.user.name
   );
