--- src/components/Views/Bodyguard/NonBodyguard/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export const BodyguaredPhaseNonBodyguardView: React.FC = () => {
-  return (
-    <React.Fragment>
-      <h2>You are not a Bodyguard</h2>
-
-      <p>The bodyguard(s) are busy saving people...</p>
-    </React.Fragment>
-  );
-};
-
-export default BodyguaredPhaseNonBodyguardView;
