--- src/components/Views/Day/Player/index.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-export const DayPlayerView: React.FC = () => (
-  <p>Talk, and suggest a lynching maybe?</p>
-);
-
-export default DayPlayerView;
