--- src/components/Views/Lobby/Player/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export const LobbyPlayerView: React.FC = () => (
-  <React.Fragment>
-    <h2>You are a Player</h2>
-    <p>Waiting for the game to start...</p>
-  </React.Fragment>
-);
-
-export default LobbyPlayerView;
