--- src/components/Views/Seer/NonSeer/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export const SeerPhaseNonSeerView: React.FC = () => {
-  return (
-    <React.Fragment>
-      <h2>You are not a Seer</h2>
-
-      <p>The seer(s) are busy seeing things...</p>
-    </React.Fragment>
-  );
-};
-
-export default SeerPhaseNonSeerView;
