+++ src/components/Views/Setup/connector/index.ts
@@ -4,6 +4,9 @@ import { IPlayer, PLAYER_ROLE } from '../../../../types/player';
 import { IState } from '../../../../store/reducers';
 
 export interface PropsFromState {
+  villageName: string;
+  lobbyId: string;
+  moderator?: IPlayer;
   players: IPlayer[];
 }
 
@@ -13,12 +16,13 @@ export const mapStateToProps = (state: IState): PropsFromState => {
   }
 
   return {
+    villageName: state.game.villageName,
+    lobbyId: state.game.lobbyId,
+    moderator: state.game.players.find(
+      (player) => player.attributes.role === PLAYER_ROLE.MODERATOR
+    ),
     players: state.game.players
-      .filter(
-        (player) =>
-          player.attributes.role !== PLAYER_ROLE.MODERATOR &&
-          player.name !== state.user.name
-      )
+      .filter((player) => player.attributes.role !== PLAYER_ROLE.MODERATOR)
       .sort((a, b) =>
         a.name.localeCompare(b.name, 'en', { sensitivity: 'base' })
       ),
