+++ src/components/Views/Setup/styles.scss
