+++ src/components/Views/TextOnly/Daytime/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import TextOnly from '..';
+
+export const DaytimeTextOnly: React.FC = () => (
+  <TextOnly
+    title="It's the day"
+    description='Have a chat, and suggest a lynching maybe? Otherwise you can always ask the moderator to sleep without killing anyone'
+  />
+);
+
+export default DaytimeTextOnly;
