+++ src/components/Views/TextOnly/NonBodyguard/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import TextOnly from '..';
+
+export const NonBodyguardTextOnly: React.FC = () => (
+  <TextOnly
+    title='You are not a Bodyguard'
+    description='The Bodyguards are busy saving people... pray its you'
+  />
+);
+
+export default NonBodyguardTextOnly;
