+++ src/components/Views/TextOnly/NonSeer/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import TextOnly from '..';
+
+export const NonSeerTextOnly: React.FC = () => (
+  <TextOnly
+    title='You are not a Seer'
+    description='The Seers are busy finding out allegiances, hang tight!'
+  />
+);
+
+export default NonSeerTextOnly;
