+++ src/components/Views/TextOnly/WaitingForStart/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import TextOnly from '..';
+
+export const WaitingForStartTextOnly: React.FC = () => (
+  <TextOnly
+    title='Waiting for the game to start'
+    description='Once everyone is ready, the Moderator will start the game, and you will find out your role! Get ready for Janky Werewolf!'
+  />
+);
+
+export default WaitingForStartTextOnly;
