--- src/components/Views/Werewolf/NonWerewolf/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export const WerewolfPhaseNonWerewolfView: React.FC = () => {
-  return (
-    <React.Fragment>
-      <h2>You are not a Werewolf</h2>
-
-      <p>The werewolves are busy deciding who to kill...</p>
-    </React.Fragment>
-  );
-};
-
-export default WerewolfPhaseNonWerewolfView;
